deleted profileController and userController and moved necesary methods to authController
