+++ routes/web.php
@@ -25,7 +25,7 @@ $router->group(['prefix' => 'api'], function ($app) use($router) {
                 $router->get('/', function () use ($router) {
                     return "api/v1";
                 });
-                $router->get('userinfo','UserController@userinfo');
+                $router->get('userinfo','AuthController@userinfo');
                 
                 $router->group(['prefix' => 'auth'], function () use($router){
                     $router->post('login','AuthController@login');
@@ -33,16 +33,13 @@ $router->group(['prefix' => 'api'], function ($app) use($router) {
                     $router->post('logout','AuthController@logout');
                     $router->post('forgetPassword','AuthController@forgetPassword');
                     $router->post('resetPassword','AuthController@resetPassword');
+                    $router->post('emailConfirmation','AuthController@emailConfirmation');
+                    
                     $router->post('changePassword','AuthController@changePassword');
                     $router->post('changeEmail','AuthController@changeEmail');
-                    $router->post('emailConfirmation','AuthController@emailConfirmation');
-                });
-
-                // Profile
-                $router->group(['prefix' => 'account'], function () use($router) {
-                    $router->post('updatePassword','ProfileController@updatePassword');
-                    $router->post('updateEmail','ProfileController@updateEmail');
-                    $router->post('updateNameAndAddress','ProfileController@updateNameAndAddress');
+                    $router->post('changeNameAndAddress','ProfileController@changeNameAndAddress');
+                    // $router->get('userinfo','UserController@userinfo');
+                    
                 });
 
                 // Profile
